+++ familyhelper-assets-impl/src/main/java/com/dwarfeng/familyhelper/assets/impl/util/FtpConstants.java
@@ -1,0 +1,16 @@
+package com.dwarfeng.familyhelper.assets.impl.util;
+
+/**
+ * FTP 常量。
+ *
+ * @author DwArFeng
+ * @since 1.0.0
+ */
+public final class FtpConstants {
+
+    public static final String PATH_ITEM_COVER = "familyhelper-assets/item-cover";
+
+    private FtpConstants() {
+        throw new IllegalStateException("禁止实例化");
+    }
+}