+++ familyhelper-assets-stack/src/main/java/com/dwarfeng/familyhelper/assets/stack/bean/dto/ItemCover.java
@@ -1,0 +1,51 @@
+package com.dwarfeng.familyhelper.assets.stack.bean.dto;
+
+import com.dwarfeng.subgrade.stack.bean.dto.Dto;
+
+import java.util.Arrays;
+
+/**
+ * 项目封面。
+ *
+ * @author DwArFeng
+ * @since 1.0.0
+ */
+public class ItemCover implements Dto {
+
+    private static final long serialVersionUID = -5892500788477351537L;
+
+    private String originName;
+    private byte[] content;
+
+    public ItemCover() {
+    }
+
+    public ItemCover(String originName, byte[] content) {
+        this.originName = originName;
+        this.content = content;
+    }
+
+    public String getOriginName() {
+        return originName;
+    }
+
+    public void setOriginName(String originName) {
+        this.originName = originName;
+    }
+
+    public byte[] getContent() {
+        return content;
+    }
+
+    public void setContent(byte[] content) {
+        this.content = content;
+    }
+
+    @Override
+    public String toString() {
+        return "ItemCover{" +
+                "originName='" + originName + '\'' +
+                ", content=" + Arrays.toString(content) +
+                '}';
+    }
+}